Rename HgConfig resources to HgGlobal
