+++ scm-plugins/scm-hg-plugin/src/main/java/sonia/scm/api/v2/resources/HgGlobalConfigAutoConfigurationResource.java
@@ -24,72 +24,48 @@
 
 package sonia.scm.api.v2.resources;
 
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import com.google.inject.Inject;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import sonia.scm.config.ConfigurationPermissions;
 import sonia.scm.repository.HgGlobalConfig;
 import sonia.scm.repository.HgRepositoryHandler;
 import sonia.scm.web.HgVndMediaType;
 import sonia.scm.web.VndMediaType;
 
-import javax.inject.Inject;
-import javax.inject.Provider;
 import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
 
-/**
- * RESTful Web Service Resource to manage the configuration of the hg plugin.
- */
-@OpenAPIDefinition(tags = {
-  @Tag(name = "Mercurial", description = "Configuration for the mercurial repository type")
-})
-@Path(HgConfigResource.HG_CONFIG_PATH_V2)
-public class HgConfigResource {
+public class HgGlobalConfigAutoConfigurationResource {
 
-  static final String HG_CONFIG_PATH_V2 = "v2/config/hg";
-  private final HgConfigDtoToHgConfigMapper dtoToConfigMapper;
-  private final HgConfigToHgConfigDtoMapper configToDtoMapper;
   private final HgRepositoryHandler repositoryHandler;
-  private final Provider<HgConfigAutoConfigurationResource> autoconfigResource;
+  private final HgGlobalConfigDtoToHgConfigMapper dtoToConfigMapper;
 
   @Inject
-  public HgConfigResource(HgConfigDtoToHgConfigMapper dtoToConfigMapper,
-                          HgConfigToHgConfigDtoMapper configToDtoMapper,
-                          HgRepositoryHandler repositoryHandler,
-                          Provider<HgConfigAutoConfigurationResource> autoconfigResource) {
+  public HgGlobalConfigAutoConfigurationResource(HgGlobalConfigDtoToHgConfigMapper dtoToConfigMapper,
+                                                 HgRepositoryHandler repositoryHandler) {
     this.dtoToConfigMapper = dtoToConfigMapper;
-    this.configToDtoMapper = configToDtoMapper;
     this.repositoryHandler = repositoryHandler;
-    this.autoconfigResource = autoconfigResource;
   }
 
   /**
-   * Returns the hg config.
+   * Sets the default hg config and installs the hg binary.
    */
-  @GET
+  @PUT
   @Path("")
-  @Produces(HgVndMediaType.CONFIG)
-  @Operation(summary = "Hg configuration", description = "Returns the global mercurial configuration.", tags = "Mercurial", operationId = "hg_get_config")
+  @Operation(summary = "Sets hg configuration and installs hg binary", description = "Sets the default mercurial config and installs the mercurial binary.", tags = "Mercurial")
   @ApiResponse(
-    responseCode = "200",
-    description = "success",
-    content = @Content(
-      mediaType = HgVndMediaType.CONFIG,
-      schema = @Schema(implementation = HgConfigDto.class)
-    )
+    responseCode = "204",
+    description = "update success"
   )
   @ApiResponse(responseCode = "401", description = "not authenticated / invalid credentials")
-  @ApiResponse(responseCode = "403", description = "not authorized, the current user does not have the \"configuration:read:hg\" privilege")
+  @ApiResponse(responseCode = "403", description = "not authorized, the current user does not have the \"configuration:write:hg\" privilege")
   @ApiResponse(
     responseCode = "500",
     description = "internal server error",
@@ -97,22 +73,12 @@ public class HgConfigResource {
       mediaType = VndMediaType.ERROR_TYPE,
       schema = @Schema(implementation = ErrorDto.class)
     ))
-  public Response get() {
-
-    ConfigurationPermissions.read(HgGlobalConfig.PERMISSION).check();
-
-    HgGlobalConfig config = repositoryHandler.getConfig();
-
-    if (config == null) {
-      config = new HgGlobalConfig();
-      repositoryHandler.setConfig(config);
-    }
-
-    return Response.ok(configToDtoMapper.map(config)).build();
+  public Response autoConfiguration() {
+    return autoConfiguration(null);
   }
 
   /**
-   * Modifies the hg config.
+   * Modifies the hg config and installs the hg binary.
    *
    * @param configDto new configuration object
    */
@@ -120,18 +86,17 @@ public class HgConfigResource {
   @Path("")
   @Consumes(HgVndMediaType.CONFIG)
   @Operation(
-    summary = "Modify hg configuration",
-    description = "Modifies the global mercurial configuration.",
+    summary = "Modifies hg configuration and installs hg binary",
+    description = "Modifies the mercurial config and installs the mercurial binary.",
     tags = "Mercurial",
-    operationId = "hg_put_config",
     requestBody = @RequestBody(
       content = @Content(
         mediaType = HgVndMediaType.CONFIG,
-        schema = @Schema(implementation = UpdateHgConfigDto.class),
+        schema = @Schema(implementation = UpdateHgGlobalConfigDto.class),
         examples = @ExampleObject(
-          name = "Overwrites current configuration with this one.",
-          value = "{\n  \"disabled\":false,\n  \"hgBinary\":\"hg\",\n  \"encoding\":\"UTF-8\",\n  \"showRevisionInId\":false,\n  \"enableHttpPostArgs\":false\n}",
-          summary = "Simple update configuration"
+          name = "Overwrites current configuration with this one and installs the mercurial binary.",
+          value = "{\n  \"disabled\":false,\n  \"hgBinary\":\"hg\",\n  \"pythonBinary\":\"python\",\n  \"pythonPath\":\"\",\n  \"encoding\":\"UTF-8\",\n  \"useOptimizedBytecode\":false,\n  \"showRevisionInId\":false,\n  \"disableHookSSLValidation\":false,\n  \"enableHttpPostArgs\":false\n}",
+          summary = "Simple update configuration and installs binary"
         )
       )
     )
@@ -149,20 +114,20 @@ public class HgConfigResource {
       mediaType = VndMediaType.ERROR_TYPE,
       schema = @Schema(implementation = ErrorDto.class)
     ))
-  public Response update(HgConfigDto configDto) {
+  public Response autoConfiguration(HgGlobalGlobalConfigDto configDto) {
 
-    HgGlobalConfig config = dtoToConfigMapper.map(configDto);
+    HgGlobalConfig config;
+
+    if (configDto != null) {
+      config = dtoToConfigMapper.map(configDto);
+    } else {
+      config = new HgGlobalConfig();
+    }
 
     ConfigurationPermissions.write(config).check();
 
-    repositoryHandler.setConfig(config);
-    repositoryHandler.storeConfig();
+    repositoryHandler.doAutoConfiguration(config);
 
     return Response.noContent().build();
-  }
-
-  @Path("auto-configuration")
-  public HgConfigAutoConfigurationResource getAutoConfigurationResource() {
-    return autoconfigResource.get();
   }
 }
