+++ scm-plugins/scm-hg-plugin/src/test/java/sonia/scm/api/v2/resources/HgGlobalConfigDtoToHgGlobalConfigMapperTest.java
@@ -24,12 +24,42 @@
 
 package sonia.scm.api.v2.resources;
 
-import org.mapstruct.Mapper;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.junit.MockitoJUnitRunner;
 import sonia.scm.repository.HgGlobalConfig;
 
-// Mapstruct does not support parameterized (i.e. non-default) constructors. Thus, we need to use field injection.
-@SuppressWarnings("squid:S3306")
-@Mapper
-public abstract class HgConfigDtoToHgConfigMapper {
-  public abstract HgGlobalConfig map(HgConfigDto dto);
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+@RunWith(MockitoJUnitRunner.class)
+public class HgGlobalConfigDtoToHgGlobalConfigMapperTest {
+
+  @InjectMocks
+  private HgGlobalConfigDtoToHgConfigMapperImpl mapper;
+
+  @Test
+  public void shouldMapFields() {
+    HgGlobalGlobalConfigDto dto = createDefaultDto();
+    HgGlobalConfig config = mapper.map(dto);
+
+    assertTrue(config.isDisabled());
+
+    assertEquals("ABC", config.getEncoding());
+    assertEquals("/etc/hg", config.getHgBinary());
+    assertTrue(config.isShowRevisionInId());
+    assertTrue(config.isEnableHttpPostArgs());
+  }
+
+  private HgGlobalGlobalConfigDto createDefaultDto() {
+    HgGlobalGlobalConfigDto configDto = new HgGlobalGlobalConfigDto();
+    configDto.setDisabled(true);
+    configDto.setEncoding("ABC");
+    configDto.setHgBinary("/etc/hg");
+    configDto.setShowRevisionInId(true);
+    configDto.setEnableHttpPostArgs(true);
+
+    return configDto;
+  }
 }
