+++ scm-plugins/scm-hg-plugin/src/test/java/sonia/scm/api/v2/resources/HgGlobalConfigInIndexResourceTest.java
@@ -49,13 +49,13 @@ public class HgGlobalConfigInIndexResourceTest {
 
   private final ObjectMapper objectMapper = new ObjectMapper();
   private final ObjectNode root = objectMapper.createObjectNode();
-  private final HgConfigInIndexResource hgConfigInIndexResource;
+  private final HgGlobalConfigInIndexResource hgGlobalConfigInIndexResource;
 
   public HgGlobalConfigInIndexResourceTest() {
     root.put("_links", objectMapper.createObjectNode());
     ScmPathInfoStore pathInfoStore = new ScmPathInfoStore();
     pathInfoStore.set(() -> URI.create("/"));
-    hgConfigInIndexResource = new HgConfigInIndexResource(Providers.of(pathInfoStore), objectMapper);
+    hgGlobalConfigInIndexResource = new HgGlobalConfigInIndexResource(Providers.of(pathInfoStore), objectMapper);
   }
 
   @Test
@@ -63,7 +63,7 @@ public class HgGlobalConfigInIndexResourceTest {
   public void admin() {
     JsonEnricherContext context = new JsonEnricherContext(URI.create("/index"), MediaType.valueOf(VndMediaType.INDEX), root);
 
-    hgConfigInIndexResource.enrich(context);
+    hgGlobalConfigInIndexResource.enrich(context);
 
     assertEquals("/v2/config/hg", root.get("_links").get("hgConfig").get("href").asText());
   }
@@ -73,7 +73,7 @@ public class HgGlobalConfigInIndexResourceTest {
   public void user() {
     JsonEnricherContext context = new JsonEnricherContext(URI.create("/index"), MediaType.valueOf(VndMediaType.INDEX), root);
 
-    hgConfigInIndexResource.enrich(context);
+    hgGlobalConfigInIndexResource.enrich(context);
 
     assertTrue(root.get("_links").iterator().hasNext());
   }
@@ -82,7 +82,7 @@ public class HgGlobalConfigInIndexResourceTest {
   public void anonymous() {
     JsonEnricherContext context = new JsonEnricherContext(URI.create("/index"), MediaType.valueOf(VndMediaType.INDEX), root);
 
-    hgConfigInIndexResource.enrich(context);
+    hgGlobalConfigInIndexResource.enrich(context);
 
     assertFalse(root.get("_links").iterator().hasNext());
   }
