+++ scm-plugins/scm-hg-plugin/src/test/java/sonia/scm/api/v2/resources/HgGlobalConfigToHgGlobalConfigDtoMapperTest.java
@@ -44,11 +44,11 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static sonia.scm.api.v2.resources.HgConfigTests.assertEqualsConfiguration;
-import static sonia.scm.api.v2.resources.HgConfigTests.createConfiguration;
+import static sonia.scm.api.v2.resources.HgGlobalConfigTestUtil.assertEqualsConfiguration;
+import static sonia.scm.api.v2.resources.HgGlobalConfigTestUtil.createConfiguration;
 
 @RunWith(MockitoJUnitRunner.class)
-public class HgConfigToHgGlobalConfigDtoMapperTest {
+public class HgGlobalConfigToHgGlobalConfigDtoMapperTest {
 
   private URI baseUri = URI.create("http://example.com/base/");
 
@@ -56,7 +56,7 @@ public class HgConfigToHgGlobalConfigDtoMapperTest {
   private ScmPathInfoStore scmPathInfoStore;
 
   @InjectMocks
-  private HgConfigToHgConfigDtoMapperImpl mapper;
+  private HgGlobalConfigToHgGlobalConfigDtoMapperImpl mapper;
 
   private final Subject subject = mock(Subject.class);
   private final ThreadState subjectThreadState = new SubjectThreadState(subject);
@@ -66,7 +66,7 @@ public class HgConfigToHgGlobalConfigDtoMapperTest {
   @Before
   public void init() {
     when(scmPathInfoStore.get().getApiRestUri()).thenReturn(baseUri);
-    expectedBaseUri = baseUri.resolve(HgConfigResource.HG_CONFIG_PATH_V2);
+    expectedBaseUri = baseUri.resolve(HgGlobalConfigResource.HG_CONFIG_PATH_V2);
     subjectThreadState.bind();
     ThreadContext.bind(subject);
   }
@@ -81,7 +81,7 @@ public class HgConfigToHgGlobalConfigDtoMapperTest {
     HgGlobalConfig config = createConfiguration();
 
     when(subject.isPermitted("configuration:write:hg")).thenReturn(true);
-    HgConfigDto dto = mapper.map(config);
+    HgGlobalGlobalConfigDto dto = mapper.map(config);
 
     assertEqualsConfiguration(dto);
 
@@ -94,7 +94,7 @@ public class HgConfigToHgGlobalConfigDtoMapperTest {
     HgGlobalConfig config = createConfiguration();
 
     when(subject.isPermitted("configuration:write:hg")).thenReturn(false);
-    HgConfigDto dto = mapper.map(config);
+    HgGlobalGlobalConfigDto dto = mapper.map(config);
 
     assertEquals(expectedBaseUri.toString(), dto.getLinks().getLinkBy("self").get().getHref());
     assertFalse(dto.getLinks().hasLink("update"));
